Add files via upload
After pip install the PyOMA module, in order to make it work when imported e.g. in Anaconda Spyder IDE, it is necessary look for the installed packages folder (C:\ProgramData\Anaconda3\Lib\site-packages\PyOMA) and paste here the file __init__.py . Now, when import PyOMA as oma in Spyder IDE it will work and load all the functions inside the PyOMA module.